Don't check for xorg/xorg-server.h, use a better error message
Checking for xorg/xorg-server.h is not necessary if pkg-config reports
that the xorg-server package is installed. We should not distrust
pkg-config without a good reason.

The check for xorg/xorg-server.h was also incorrect, as it didn't use
the include path returned by pkg-config for xorg-server. Thus the check
would fail if the header was not in the default include path.

Suggest installing xorg-x11-server-devel (RedHat) in addition to
xserver-xorg-dev (Debian) and xorg-x11-server-sdk (openSUSE).