Add --enable-strict-locations options, use it in "make distcheck"
When --enable-strict-locations is used, the location of Xorg modules is
not autodetected. Configuration files are installed under sysconfdir.

Don't make any logic depend on the presence of the --prefix option, it's
not an expected side effect.

Use --enable-strict-locations in "make distcheck", it's needed to make
sure all files are installed under prefix.